Revert "Merge tag 'android-4.4.3_r1.1' into HEAD"
This reverts commit 9fcf3a7532582ab4273d3ebbef065da37a1225c4.

+++ include/utils/SkMeshUtils.h
@@ -27,14 +27,14 @@ public:
     bool init(SkPoint tex[], uint16_t indices[],
               int texW, int texH, int rows, int cols);
 
-    int             indexCount() const { return fIndexCount; }
+    size_t          indexCount() const { return fIndexCount; }
     const uint16_t* indices() const { return fIndices; }
 
     size_t          texCount() const { return fTexCount; }
     const SkPoint*  tex() const { return fTex; }
 
 private:
-    int         fIndexCount, fTexCount;
+    size_t      fIndexCount, fTexCount;
     SkPoint*    fTex;
     uint16_t*   fIndices;
     void*       fStorage; // may be null
